Removed warning about ssl verification in the sinatra app
Apparently rest-client which is used in the actual library doesn't verify SSL
by default either. No point in warning about the sinatra utility app when it's
true for the whole library (although the real answer is to just fix the
problem in both places; adding an issue so I don't forget about it).